third_party/WebKit/LayoutTests/editing/pasteboard/paste_plain_text_to_newlines.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_not_equals(window.internals, undefined,
        'This test requires window.internals to access clipboard');

    assert_selection(
        '<div contenteditable><b>|<br></b></div>',
        selection => {
            selection.setClipboardData('abc');
            selection.document.execCommand('pasteAndMatchStyle');
        },
        '<div contenteditable><b>abc|</b></div>',
        'pasteAndMatchStyle');
});
</script>
